+++ bindings/vue/src/components/VOnsNavigator.vue
@@ -46,6 +46,11 @@
           && this.$ons._ons._util.getTopPage(page._contentElement.children[0]);
         return nextPage ? this._findScrollPage(nextPage) : page;
       },
+      _setPagesVisibility(start, end, property, value) {
+        for (let i = start; i < end; i++) {
+          this.$children[i].$el.style[property] = value;
+        }
+      },
       _reattachPage(pageElement, position = null, restoreScroll) {
         this.$el.insertBefore(pageElement, position);
         restoreScroll instanceof Function && restoreScroll();
@@ -65,14 +70,12 @@
             isReattached = true;
             lastLength--;
           }
-
-          currentTopPage.style.visibility = 'hidden'; // Avoid flickerings
-          for (let i = lastLength; i < currentLength - 1; i++) {
-            this.$children[i].$el.style.display = 'none'; // Avoid flickerings and perf
-          }
+          this._setPagesVisibility(lastLength, currentLength, 'visibility', 'hidden');
 
           return this.$el._pushPage({ ...this.options, leavePage: lastTopPage })
             .then(() => {
+              this._setPagesVisibility(lastLength, currentLength, 'visibility', '');
+              this._setPagesVisibility(lastLength, currentLength - 1, 'display', 'none');
               if (isReattached) {
                 this._redetachPage(lastTopPage);
               }
@@ -86,7 +89,7 @@
         }
 
         // Replace page
-        currentTopPage.style.visibility = 'hidden'; // Avoid flickerings
+        currentTopPage.style.visibility = 'hidden';
         this._reattachPage(lastTopPage, currentTopPage, restoreScroll);
         return this.$el._pushPage({ ...this.options, _replacePage: true }).then(() => this._redetachPage(lastTopPage));
       },