+++ core/src/ons/platform.js
@@ -109,7 +109,9 @@ class Platform {
   isIPhoneX() {
     // iPhone 8 and iPhone X have a same user agent. We cannot avoid using window.screen.
     // This works well both in iOS Safari and (UI|WK)WebView of iPhone X.
-    return this.isIPhone() && window.screen.width === 375 && window.screen.height === 812;
+    return this.isIPhone() &&
+      (window.screen.width === 375 && window.screen.height === 812
+    || window.screen.width === 812 && window.screen.height === 375);
   }
 
   /**