+++ core/src/setup.js
@@ -1,3 +1,4 @@
+import { FastClick } from '@onsenui/fastclick';
 import './ons/platform'; // This file must be loaded before Custom Elements polyfills.
 import './polyfills/index.js';
 import './vendor/index.js';